requirements.txt
py-cord==2.4.1
requests==2.31.0
python-dotenv==1.0.0
PyNaCl==1.5.0
pytz==2023.3.post1
yt-dlp==2023.10.13
dpytest==0.7.0
cassandra-driver==3.28.0
axiom-py==0.2.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-exporter-otlp==1.21.0